requirements.txt
chromadb>=0.5.0
pdfplumber>=0.11.0
langchain-text-splitters>=0.2.0
google-genai>=1.0.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
streamlit>=1.39.0
requests>=2.32.0
